refactor(scripts): dedupe mcp code-mode fixture
